+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Contributor to this version: Baptiste Hamon (@baptistehamon).
   - Grapevines
     - Pinot Noir ([GH#35](https://github.com/baptistehamon/nzlusdb/issues/35), [PR#37](https://github.com/baptistehamon/nzlusdb/pull/37))
     - Sauvignon Blanc ([GH#35](https://github.com/baptistehamon/nzlusdb/issues/35), [PR#37](https://github.com/baptistehamon/nzlusdb/pull/37))
+  - Early and late wheat ([GH#38](https://github.com/baptistehamon/nzlusdb/issues/38), [PR#42](https://github.com/baptistehamon/nzlusdb/pull/42))
 
 ### Internal Changes
 - A preprocessing step has been added enabling to preprocess criteria indicators before running the LSA ([GH9#](https://github.com/baptistehamon/nzlusdb/issues/9)).
@@ -36,6 +37,8 @@ Contributor to this version: Baptiste Hamon (@baptistehamon).
 - The `cracking_survival` index and indicator have been added to support the computation of cherry climate indicators ([PR#34](https://github.com/baptistehamon/nzlusdb/pull/34)).
 - The `chilling_hours` indicator has been modified to include both upper and lower temperature thresholds ([PR#37](https://github.com/baptistehamon/nzlusdb/pull/37)).
 - The `heat_survival` index and indicator have been added to support the computation of grapevine climate indicators ([PR#37](https://github.com/baptistehamon/nzlusdb/pull/37)).
+- The criteria indicator preprocessing functionality has been used to update allowing custom functions ([PR#42](https://github.com/baptistehamon/nzlusdb/pull/42)).
+- The `decode_timedelta` of `xr.open_dataarray` in `nzlusdb.core.landuse.LandUse._load_indicator` has been set to `False` ([PR#42](https://github.com/baptistehamon/nzlusdb/pull/42)).
 
 ### Bug Fixes
 - Fix label error for projected suitability changes summary figures ([GH#11](https://github.com/baptistehamon/nzlusdb/issues/11), [PR#19](https://github.com/baptistehamon/nzlusdb/pull/19)).